chore: rename CopyTooltipText
